斐波那契数列

+++ rp_09_Fibonacci.c
@@ -2,23 +2,22 @@
 // Created by C2243829 on 2021/2/17.
 //
 #include "stdio.h"
-int main(){
+
+int main() {
     int n;
     int i;
-    long int f3, f2, f1;
+    int f3, f2, f1;
     f2 = 1;
     f1 = 1;
     printf("请输入你需要求的项的序列:");
     scanf("%d", &n);
-    if (1 == n){
+    if (1 == n) {
         f3 = 1;
-    }
-    else if (2 == n){
+    } else if (2 == n) {
         f3 = 1;
-    }
-    else{
+    } else {
 
-        for (i = 3;  i <= n; ++i) {
+        for (i = 3; i <= n; ++i) {
             f3 = f2 + f1;
             f1 = f2;
             f2 = f3;
